double -> float hyperbolic + HyperbolicTranslationXW

+++ NEEngine/inc/HyperbolicMath.h
@@ -1,8 +1,9 @@
 #pragma once
 
-DirectX::SimpleMath::Matrix HyperbolicTranslationZW(double d);
-DirectX::SimpleMath::Matrix HyperbolicTranslationXW(double d);
-DirectX::SimpleMath::Matrix HyperbolicTranslationYW(double d);
-DirectX::SimpleMath::Matrix HyperbolicRotationYZ(double d);
-DirectX::SimpleMath::Matrix HyperbolicRotationXZ(double d);
-DirectX::SimpleMath::Matrix HyperbolicRotationXY(double d);
+DirectX::SimpleMath::Matrix HyperbolicTranslationZW(float d);
+DirectX::SimpleMath::Matrix HyperbolicTranslationXW(float d);
+DirectX::SimpleMath::Matrix HyperbolicTranslationYW(float d);
+
+DirectX::SimpleMath::Matrix HyperbolicRotationYZ(float d);
+DirectX::SimpleMath::Matrix HyperbolicRotationXZ(float d);
+DirectX::SimpleMath::Matrix HyperbolicRotationXY(float d);
